Fix : mypage/favorite Suspense 미적용으로 인한 오류 해결
