+++ next.config.mjs
@@ -1,7 +1,7 @@
 import fs from 'fs';
 import path from 'path';
 
-const envMode = process.env.ENV_MODE || 'default';
+const envMode = process.env.ENV_MODE;
 const envFile = path.resolve(process.cwd(), `.env.${envMode}`);
 
 if (fs.existsSync(envFile)) {
